refactor: `torus-sdk-ts` chain filesystem
