+++ packages/torus-sdk-ts/src/agent/agent.ts
@@ -10,10 +10,10 @@ import { z } from "zod";
 
 import type { SS58Address } from "@torus-network/sdk/types";
 
-import { queryNamespacePermissions } from "../chain/permission0.js";
-import { queryAgents } from "../chain/torus0/agents.js";
-import { connectToChainRpc } from "../utils/index.js";
+import { queryNamespacePermissions } from "../chain/permission0/permission0-storage.js";
+import { queryAgents } from "../chain/torus0/torus0-storage.js";
 import { validateNamespacePath } from "../types/namespace/namespace-path.js";
+import { connectToChainRpc } from "../utils/index.js";
 import type { Helpers } from "./helpers.js";
 import { checkTransaction } from "./helpers.js";
 import type { AuthTokenResult } from "./utils.js";
