+++ packages/torus-sdk-ts/src/chain/emission0/emission0-storage.ts
@@ -1,12 +1,9 @@
-import type { ApiPromise } from "@polkadot/api";
-import type { KeyringPair } from "@polkadot/keyring/types";
 import type { Percent } from "@polkadot/types/interfaces";
 
 import { tryAsync, trySync } from "@torus-network/torus-utils/try-catch";
 
-import type { SS58Address } from "../types/address.js";
-import { sb_bigint } from "../types/index.js";
-import type { Api } from "./common/index.js";
+import { sb_bigint } from "../../types/index.js";
+import type { Api } from "../common/fees.js";
 
 export async function queryRecyclingPercentage(api: Api): Promise<Percent> {
   const [queryError, recyclingPercentage] = await tryAsync(
@@ -47,31 +44,4 @@ export function queryBlockEmission(api: Api): bigint {
   }
 
   return emission;
-}
-
-// ==== Weights ====
-
-export async function setChainWeights(
-  api: ApiPromise,
-  keypair: KeyringPair,
-  weights: [SS58Address, number][],
-) {
-  const [createTxError, tx] = trySync(() =>
-    api.tx.emission0.setWeights(weights),
-  );
-  if (createTxError !== undefined) {
-    console.error("Error creating set weights transaction:", createTxError);
-    throw createTxError;
-  }
-
-  const [signError, signedTx] = await tryAsync(tx.signAndSend(keypair));
-  if (signError !== undefined) {
-    console.error(
-      "Error signing and sending set weights transaction:",
-      signError,
-    );
-    throw signError;
-  }
-
-  return signedTx;
 }
