+++ packages/torus-sdk-ts/src/chain/permission0/permission0-extrinsics.ts
@@ -1,0 +1,180 @@
+import type { ApiPromise } from "@polkadot/api";
+import type { Bytes, Option, u16 } from "@polkadot/types";
+import { BTreeMap, BTreeSet } from "@polkadot/types";
+import type { AccountId32, H256, Percent } from "@polkadot/types/interfaces";
+
+import type { SS58Address } from "../../types/index.js";
+import type {
+  DistributionControl,
+  EmissionAllocation,
+  EnforcementAuthority,
+  PermissionDuration,
+  PermissionId,
+  RevocationTerms,
+  StreamId,
+} from "./permission0-types.js";
+
+export interface DelegateNamespacePermission {
+  api: ApiPromise;
+  recipient: SS58Address;
+  paths: Map<H256 | null, string[]>;
+  duration: PermissionDuration;
+  revocation: RevocationTerms;
+  instances: number;
+}
+
+/**
+ * Delegate a permission over namespaces
+ */
+export function delegateNamespacePermission({
+  api,
+  recipient,
+  paths,
+  duration,
+  revocation,
+  instances,
+}: DelegateNamespacePermission) {
+  // Convert the paths map to BTreeMap with BTreeSet values
+  const pathsMap = new BTreeMap<Option<H256>, BTreeSet<Bytes>>(
+    api.registry,
+    "Option<H256>",
+    "BTreeSet<Bytes>",
+    new Map(),
+  );
+
+  for (const [parent, pathList] of paths.entries()) {
+    const btreeSet = new BTreeSet<Bytes>(api.registry, "Bytes", pathList);
+    // Convert null to Option<H256>
+    const optionParent =
+      parent === null
+        ? api.createType("Option<H256>", null)
+        : api.createType("Option<H256>", parent);
+    pathsMap.set(optionParent, btreeSet);
+  }
+
+  return api.tx.permission0.delegateNamespacePermission(
+    recipient,
+    pathsMap,
+    duration,
+    revocation,
+    instances,
+  );
+}
+
+/**
+ * Delegate an emission permission to a recipient
+ */
+export interface DelegateEmissionPermission {
+  api: ApiPromise;
+  recipient: string;
+  allocation: EmissionAllocation;
+  targets: [SS58Address, number][];
+  distribution: DistributionControl;
+  duration: PermissionDuration;
+  revocation: RevocationTerms;
+  enforcement: EnforcementAuthority;
+}
+
+/**
+ * TODO: test
+ * TODO: docs
+ */
+export function delegateEmissionPermission({
+  api,
+  recipient,
+  allocation,
+  targets,
+  distribution,
+  duration,
+  revocation,
+  enforcement,
+}: DelegateEmissionPermission) {
+  const targetsMap = new Map(targets);
+
+  const targetsMap_ = new BTreeMap<AccountId32, u16>(
+    api.registry,
+    "AccountId32",
+    "u32",
+    targetsMap,
+  );
+
+  return api.tx.permission0.delegateEmissionPermission(
+    recipient,
+    allocation,
+    targetsMap_,
+    distribution,
+    duration,
+    revocation,
+    enforcement,
+  );
+}
+
+export function togglePermission(
+  api: ApiPromise,
+  permissionId: PermissionId,
+  enable: boolean,
+) {
+  return api.tx.permission0.togglePermissionAccumulation(permissionId, enable);
+}
+
+export interface UpdateEmissionPermission {
+  api: ApiPromise;
+  permissionId: PermissionId;
+  newTargets?: [SS58Address, number][];
+  newStreams?: Map<StreamId, number>;
+  newDistributionControl?: DistributionControl;
+}
+
+/**
+  If you call as a recipient:
+  you can only provide the new_targets,
+  whenever you want, no limits. if the recipient sends
+  new_streams/new_distribution_control, the extrinsic fails.
+
+  If you call as a delegator:
+  you can send all the values, 
+  but only if the revocation term: is RevocableByDelegator
+  is RevocableAfter(N) and CurrentBlock > N
+  think of it as the revocation term defining whether
+  the delegator can modify the contract without
+  breaching the "terms of service"
+ */
+export function updateEmissionPermission({
+  api,
+  permissionId,
+  newTargets,
+  newStreams,
+  newDistributionControl,
+}: UpdateEmissionPermission) {
+  const targetsMap = newTargets
+    ? new BTreeMap<AccountId32, u16>(
+        api.registry,
+        "AccountId32",
+        "u16",
+        new Map(newTargets),
+      )
+    : new BTreeMap<AccountId32, u16>(
+        api.registry,
+        "AccountId32",
+        "u16",
+        new Map(),
+      );
+
+  const streamsMap = newStreams
+    ? new BTreeMap<H256, Percent>(api.registry, "H256", "Percent", newStreams)
+    : null;
+
+  return api.tx.permission0.updateEmissionPermission(
+    permissionId,
+    targetsMap,
+    streamsMap,
+    newDistributionControl ?? null,
+  );
+}
+
+/**
+ * Revoke a permission. The caller must met revocation constraints or be a root key.
+ **/
+export function revokePermission(api: ApiPromise, permissionId: PermissionId) {
+  return api.tx.permission0.revokePermission(permissionId);
+}
