--- packages/torus-sdk-ts/src/chain/torus0/index.ts
@@ -1,3 +1,0 @@
-export * from "./agents.js";
-export * from "./namespace.js";
-export * from "./staking.js";
